Add climb levels and color to nodes
