+++ games/config/2026/Nodes.svelte
@@ -104,12 +104,12 @@
       <span class="indicator-item indicator-left indicator-left indicator-start badge badge-primary">
         {gameLevelMap[level]}
       </span>
-      <span class="badge badge-accent ">
+      <span class="badge badge-success">
         {$gameData[gameMode + gameLevelMap[level] + "Fuel" + type]}
       </span>
     {:else}
       {#if level == 1 || level == 2}
-        <span class="badge badge-accent ">
+        <span class="badge badge-error">
           {$gameData[gameMode + gameLevelMap[level] + "Fuel" + type]}
         </span>
       {/if}
